test: adapt test
